+++ Engine/AppInstance.cpp
@@ -782,6 +782,8 @@ AppInstance::loadPythonScript(const QFileInfo& file)
 
         getProject()->forceComputeInputDependentDataOnAllTrees();
     } else {
+        PythonGILLocker pgl;
+
         PyRun_SimpleString( content.toStdString().c_str() );
 
         PyObject* mainModule = NATRON_PYTHON_NAMESPACE::getMainModule();